feat: evaluation layer (boundary agreement + dwell times)
- states_to_boundaries: decoded states -> boundary times
- boundary_agreement: order-independent two-pointer matching within
  tolerance, returns precision/recall/F + raw counts
- dwell_times: per-state run-length distribution
- Add Sleep ground-truth annotations (6 by-ear section boundaries)

Standard-HMM baseline on Sleep: recall 0.5, precision 0.086 (35 predicted
vs 6 true). Dwell analysis shows state 0 collapsed into flicker (median
1.5s, 8/14 dwells under 2s) while state 1 is a healthy long-duration
regime. Quantifies the over-segmentation the HSMM extension must fix.

+++ src/song_state_observer/preprocess.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 def standardise(X, eps = 1e-8):
-    X = np.asarray(X)
+    X = np.asarray(X, dtype = float)
 
     if  X.ndim != 2:
         raise ValueError("features must have shape (T, D).")
